+++ malachite-nz/src/natural/conversion/from_floating_point.rs
@@ -69,7 +69,7 @@ macro_rules! float_impls {
                     let value_negative = value < 0.0;
                     mantissa.set_bit($f::MANTISSA_WIDTH);
                     let n = Natural::from(mantissa).shl_round(
-                        i32::exact_from(exponent) + $f::MIN_EXPONENT - 1,
+                        i64::exact_from(exponent) + $f::MIN_EXPONENT - 1,
                         if value_negative { -rm } else { rm },
                     );
                     if value_negative && n != 0 {
@@ -165,7 +165,7 @@ macro_rules! float_impls {
                 } else {
                     let (mut mantissa, exponent) = value.to_adjusted_mantissa_and_exponent();
                     mantissa.set_bit($f::MANTISSA_WIDTH);
-                    let exponent = i32::exact_from(exponent) + $f::MIN_EXPONENT - 1;
+                    let exponent = i64::exact_from(exponent) + $f::MIN_EXPONENT - 1;
                     if exponent >= 0
                         || mantissa.divisible_by_power_of_two(u64::wrapping_from(-exponent))
                     {
@@ -219,7 +219,7 @@ macro_rules! float_impls {
                 } else {
                     let (mut mantissa, exponent) = value.to_adjusted_mantissa_and_exponent();
                     mantissa.set_bit($f::MANTISSA_WIDTH);
-                    let exponent = i32::exact_from(exponent) + $f::MIN_EXPONENT - 1;
+                    let exponent = i64::exact_from(exponent) + $f::MIN_EXPONENT - 1;
                     exponent >= 0
                         || mantissa.divisible_by_power_of_two(u64::wrapping_from(-exponent))
                 }
